Extract master/slave code into separate module
