+++ lib/active_record_shards/connection_switcher.rb
@@ -44,76 +44,6 @@ module ActiveRecordShards
       switch_connection(old_options)
     end
 
-    def on_slave_if(condition, &block)
-      condition ? on_slave(&block) : yield
-    end
-
-    def on_slave_unless(condition, &block)
-      on_slave_if(!condition, &block)
-    end
-
-    def on_master_if(condition, &block)
-      condition ? on_master(&block) : yield
-    end
-
-    def on_master_unless(condition, &block)
-      on_master_if(!condition, &block)
-    end
-
-    def on_master_or_slave(which, &block)
-      if block_given?
-        on_cx_switch_block(which, &block)
-      else
-        MasterSlaveProxy.new(self, which)
-      end
-    end
-
-    # Executes queries using the slave database. Fails over to master if no slave is found.
-    # if you want to execute a block of code on the slave you can go:
-    #   Account.on_slave do
-    #     Account.first
-    #   end
-    # the first account will be found on the slave DB
-    #
-    # For one-liners you can simply do
-    #   Account.on_slave.first
-    def on_slave(&block)
-      on_master_or_slave(:slave, &block)
-    end
-
-    def on_master(&block)
-      on_master_or_slave(:master, &block)
-    end
-
-    def force_cx_switch_slave_block
-      old_options = current_shard_selection.options
-      switch_connection(slave: true)
-      yield
-    ensure
-      switch_connection(old_options) if old_options
-    end
-
-    def on_cx_switch_block(which, &block)
-      @disallow_slave ||= 0
-      @disallow_slave += 1 if which == :master
-
-      switch_to_slave = @disallow_slave.zero?
-      old_options = current_shard_selection.options
-
-      switch_connection(slave: switch_to_slave)
-
-      # we avoid_readonly_scope to prevent some stack overflow problems, like when
-      # .columns calls .with_scope which calls .columns and onward, endlessly.
-      if self == ActiveRecord::Base || !switch_to_slave
-        yield
-      else
-        readonly.scoping(&block)
-      end
-    ensure
-      @disallow_slave -= 1 if which == :master
-      switch_connection(old_options) if old_options
-    end
-
     def connection_specification_name
       name = connection_resolver.resolve_connection_name(connection_config)
 
@@ -139,10 +69,6 @@ module ActiveRecordShards
 
     def supports_sharding?
       shard_names.any?
-    end
-
-    def on_slave?
-      current_shard_selection.on_slave?
     end
 
     def current_shard_selection
